+++ src/gsy_e/gsy_e_core/global_objects_singleton.py
@@ -17,10 +17,12 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 from gsy_e.gsy_e_core.user_profile_handler import ProfilesHandler
 from gsy_e.gsy_e_core.global_stats import ExternalConnectionGlobalStatistics
-from gsy_e.gsy_e_core.util import FutureMarketCounter
+from gsy_e.gsy_e_core.market_utils import FutureMarketCounter
 
 
 class GlobalObjects:
+    """Collection of globally accessible objects"""
+
     profiles_handler = ProfilesHandler()
     external_global_stats = ExternalConnectionGlobalStatistics()
     future_market_counter = FutureMarketCounter()
