+++ src/gsy_e/gsy_e_core/global_stats.py
@@ -16,8 +16,9 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 import gsy_e.constants
+from gsy_e.gsy_e_core.market_utils import ExternalTickCounter
 from gsy_e.gsy_e_core.util import (find_object_of_same_weekday_and_time,
-                                   get_market_maker_rate_from_config, ExternalTickCounter)
+                                   get_market_maker_rate_from_config)
 
 
 class ExternalConnectionGlobalStatistics:
