+++ src/gsy_e/gsy_e_core/util.py
@@ -29,12 +29,11 @@ from typing import TYPE_CHECKING
 
 from click.types import ParamType
 from gsy_framework.constants_limits import ConstSettings, GlobalConfig, RangeLimit
-from gsy_framework.enums import BidOfferMatchAlgoEnum
 from gsy_framework.exceptions import GSyException
 from gsy_framework.utils import (
     area_name_from_area_or_ma_name, iterate_over_all_modules, str_to_pendulum_datetime,
     find_object_of_same_weekday_and_time)
-from pendulum import duration, from_format, instance, DateTime
+from pendulum import duration, from_format, instance
 from rex import rex
 
 import gsy_e
@@ -482,65 +481,10 @@ def get_simulation_queue_name():
             if listen_to_cn else ConstSettings.GeneralSettings.SIM_JOB_QUEUE_NAME)
 
 
-class ExternalTickCounter:
-    """External tick counter."""
-
-    def __init__(self, ticks_per_slot: int, dispatch_frequency_percent: int):
-        self._dispatch_tick_frequency = int(
-            ticks_per_slot *
-            (dispatch_frequency_percent / 100)
-        )
-
-    def is_it_time_for_external_tick(self, current_tick_in_slot: int) -> bool:
-        """Boolean return if time for external tick."""
-        return current_tick_in_slot % self._dispatch_tick_frequency == 0
-
-
 def should_read_profile_from_db(profile_uuid):
     """Boolean return if profile to be read from DB."""
     return profile_uuid is not None and gsy_e.constants.CONNECT_TO_PROFILES_DB
 
 
-def is_external_matching_enabled():
-    """Checks if the bid offer match type is set to external
-    Returns True if both are matched
-    """
-    return (ConstSettings.MASettings.BID_OFFER_MATCH_TYPE ==
-            BidOfferMatchAlgoEnum.EXTERNAL.value)
-
-
 class StrategyProfileConfigurationException(Exception):
     """Exception raised when neither a profile nor a profile_uuid are provided for a strategy."""
-
-
-def is_time_slot_in_past_markets(time_slot: DateTime, current_time_slot: DateTime):
-    """Checks if the time_slot should be in the area.past_markets."""
-    if ConstSettings.SettlementMarketSettings.ENABLE_SETTLEMENT_MARKETS:
-        return (time_slot < current_time_slot.subtract(
-            hours=ConstSettings.SettlementMarketSettings.MAX_AGE_SETTLEMENT_MARKET_HOURS))
-    return time_slot < current_time_slot
-
-
-class FutureMarketCounter:
-    """Hold a time counter for the future market.
-
-    In the future market, we only want to clear in a predefined interval.
-    """
-    def __init__(self):
-        self._last_time_dispatched = None
-
-    def is_time_for_clearing(self, current_time: DateTime) -> bool:
-        """Compare current time with the latest time clearing was dispatched.
-
-        Returns True if the FUTURE_MARKET_CLEARING_INTERVAL_MINUTES has
-        already passed since the last dispatch time.
-        """
-        if not self._last_time_dispatched:
-            self._last_time_dispatched = current_time
-            return True
-        duration_in_min = (current_time - self._last_time_dispatched).minutes
-        if (duration_in_min >=
-                ConstSettings.FutureMarketSettings.FUTURE_MARKET_CLEARING_INTERVAL_MINUTES):
-            self._last_time_dispatched = current_time
-            return True
-        return False
